file uploaded

+++ controllers/posts_controller.js
@@ -26,12 +26,20 @@ module.exports.create=async function(req,res){
 module.exports.destroyed=async function(req, res){
     try {
        
-        const post=await Post.findById(req.params.id)
+        let post=await Post.findById(req.params.id)
         // .id means converting the object id into string
         if (post.user == req.user.id) {
-            await post.deleteOne(); // Delete the post
+             post.deleteOne(); // Delete the post
+            if(req.xhr){
+                return res.status(200).json({
+                    data: {
+                        post_id:req.params.id
+                    },
+                    messege: "Post Deleted"
+                })
+            }
            
-            await Comment.deleteMany({ post: req.params.id }); // Delete associated comments
+             Comment.deleteMany({ post: req.params.id }); // Delete associated comments
              req.flash('success','Post and associated comments deleted')
             return res.redirect('back');
         } else {
